Full Moveset Viewing
I have now added the functionality of viewing entire move sets! These moves have always been searchable on my site, but now you can straight up see them all in a list. You don't have to go through the wiki anymore!

Access this feature by clicking a Pokemon's name in the list.

The info is color coded to show you the type, damage category, and number of targets. Look at the image I posted. You'll get used to it.

Unlike the wiki, all the moves are clickable! Click on a move to see a pop up with additional information, including properties/tags.

You can use the left and right arrows to switch which Pokemon you are looking at the moveset for. This respects any filters that you have in the main list, which scrolls down automatically if you change which moveset you are viewing.